requirements.txt
shapely==2.1.2
python-slugify==8.0.4
openai==2.6.1
